Revert "refactor(rust): split monitoring_handlers en sous-module (helpers/types/monitoring/security/dns)"
This reverts commit ab6c199a0aeef13481b743d3dd08d42154ce3a31.
