+++ src/bin/email_api_dir/monitoring_handlers.rs
@@ -1,3 +1,7 @@
+// monitoring_handlers.rs — extracted from email_api_dir/main.rs Sprint 2
+// Handlers: api_monitoring_*, api_security_*
+// All shared types/imports are redeclared here so this module compiles independently.
+
 use actix_web::{web, HttpResponse};
 use chrono::Utc;
 use futures_util::stream;
@@ -10,8 +14,102 @@ use simple_smtp_server::security;
 use std::sync::Arc;
 use tokio::sync::broadcast;
 
-use super::helpers::*;
-use super::types::*;
+// ─── Shared helpers ────────────────────────────────────────────────────────
+
+pub(crate) fn parse_window(s: &str) -> chrono::Duration {
+    let s = s.trim();
+    if let Some(n) = s.strip_suffix('m').and_then(|n| n.parse::<i64>().ok()) {
+        chrono::Duration::minutes(n)
+    } else if let Some(n) = s.strip_suffix('h').and_then(|n| n.parse::<i64>().ok()) {
+        chrono::Duration::hours(n)
+    } else if let Some(n) = s.strip_suffix('d').and_then(|n| n.parse::<i64>().ok()) {
+        chrono::Duration::days(n)
+    } else {
+        chrono::Duration::minutes(15)
+    }
+}
+
+pub(crate) fn since_str(window: &str) -> String {
+    let dur = parse_window(window);
+    (Utc::now() - dur).to_rfc3339()
+}
+
+pub(crate) fn default_monitoring_window() -> String {
+    "15m".into()
+}
+
+pub(crate) fn default_window() -> String {
+    "1h".into()
+}
+
+pub(crate) fn default_mon_page() -> u32 {
+    1
+}
+pub(crate) fn default_mon_page_size() -> u32 {
+    50
+}
+
+pub(crate) fn one() -> u32 {
+    1
+}
+pub(crate) fn twenty() -> u32 {
+    20
+}
+
+// ─── Query types ───────────────────────────────────────────────────────────
+
+#[derive(Deserialize)]
+pub(crate) struct MonitoringWindowQuery {
+    #[serde(default = "default_monitoring_window")]
+    pub window: String,
+}
+
+#[derive(Deserialize)]
+pub(crate) struct MonitoringEventsQuery {
+    pub status: Option<String>,
+    pub from: Option<String>,
+    pub to: Option<String>,
+    pub provider: Option<String>,
+    pub country: Option<String>,
+    pub since: Option<String>,
+    pub until: Option<String>,
+    pub message_id: Option<String>,
+    #[serde(default = "default_mon_page")]
+    pub page: u32,
+    #[serde(default = "default_mon_page_size")]
+    pub page_size: u32,
+}
+
+#[derive(Deserialize)]
+pub(crate) struct MonitoringLiveQuery {
+    pub message_id: Option<String>,
+}
+
+#[derive(Deserialize)]
+pub(crate) struct AdminWindowQuery {
+    #[serde(default = "default_window")]
+    pub window: String,
+}
+
+#[derive(serde::Deserialize)]
+pub(crate) struct SecurityAlertsQuery {
+    #[serde(default = "default_window")]
+    pub window: String,
+    pub severity: Option<String>,
+    pub tenant_id: Option<String>,
+}
+
+#[derive(serde::Deserialize)]
+pub(crate) struct SecurityIncidentsQuery {
+    #[serde(default = "one")]
+    pub page: u32,
+    #[serde(default = "twenty")]
+    pub page_size: u32,
+    pub tenant_id: Option<String>,
+    pub severity: Option<String>,
+}
+
+// ─── Monitoring handlers ────────────────────────────────────────────────────
 
 /// GET /api/monitoring/summary?window=15m
 pub(crate) async fn api_monitoring_summary(
@@ -270,4 +368,119 @@ pub(crate) async fn api_monitoring_alerts_active(
     let config = AlertConfig::default();
     let alerts = monitoring::alerts::evaluate_alerts(&mongo, window_minutes, &config).await;
     HttpResponse::Ok().json(serde_json::json!({ "window": query.window, "alert_count": alerts.len(), "alerts": alerts }))
+}
+
+// ─── Security handlers ──────────────────────────────────────────────────────
+
+pub(crate) async fn api_security_alerts_active(
+    query: web::Query<SecurityAlertsQuery>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl actix_web::Responder {
+    use simple_smtp_server::security::audit;
+    let mut alerts = audit::query_active_alerts(&mongo, 100).await;
+    if let Some(ref sev) = query.severity {
+        let sev_lc = sev.to_lowercase();
+        alerts.retain(|a| format!("{:?}", a.severity).to_lowercase() == sev_lc);
+    }
+    if let Some(ref tid) = query.tenant_id {
+        alerts.retain(|a| a.tenant_id.as_deref() == Some(tid.as_str()));
+    }
+    HttpResponse::Ok().json(serde_json::json!({ "window": query.window, "alert_count": alerts.len(), "alerts": alerts }))
+}
+
+pub(crate) async fn api_security_incidents(
+    query: web::Query<SecurityIncidentsQuery>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl actix_web::Responder {
+    use simple_smtp_server::security::audit;
+    let mut filter = doc! {};
+    if let Some(ref tid) = query.tenant_id { filter.insert("tenant_id", tid.as_str()); }
+    if let Some(ref sev) = query.severity  { filter.insert("severity",  sev.as_str()); }
+    let alerts = audit::query_alerts(&mongo, filter, query.page, query.page_size).await;
+    HttpResponse::Ok().json(serde_json::json!({ "page": query.page, "page_size": query.page_size, "count": alerts.len(), "alerts": alerts }))
+}
+
+pub(crate) async fn api_security_tenant_status(
+    path: web::Path<String>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl actix_web::Responder {
+    let tenant_id = path.into_inner();
+    let db = std::env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
+    let coll = mongo.database(&db).collection::<mongodb::bson::Document>("tenant_state");
+    match coll.find_one(doc! { "tenant_id": &tenant_id }).await {
+        Ok(Some(doc)) => HttpResponse::Ok().json(serde_json::json!({ "tenant_id": tenant_id, "state": doc })),
+        Ok(None) => HttpResponse::Ok().json(serde_json::json!({ "tenant_id": tenant_id, "state": null })),
+        Err(e) => HttpResponse::InternalServerError().json(serde_json::json!({ "error": e.to_string() })),
+    }
+}
+
+pub(crate) async fn api_security_rollback(
+    path: web::Path<String>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl actix_web::Responder {
+    use simple_smtp_server::security::remediation;
+    let alert_id = path.into_inner();
+    match remediation::rollback_remediation(&mongo, &alert_id).await {
+        Ok(()) => HttpResponse::Ok().json(serde_json::json!({ "rolled_back": true, "alert_id": alert_id })),
+        Err(e) => HttpResponse::BadRequest().json(serde_json::json!({ "error": e })),
+    }
+}
+
+pub(crate) async fn api_security_live(mongo: web::Data<Arc<mongodb::Client>>) -> impl actix_web::Responder {
+    let rx = match security::get_bus() {
+        Some(tx) => tx.subscribe(),
+        None => return HttpResponse::ServiceUnavailable().body("security bus unavailable"),
+    };
+    let stream = stream::unfold(
+        (rx, tokio::time::interval(std::time::Duration::from_secs(15)), mongo.clone()),
+        |(mut rx, mut hb, mongo)| async move {
+            tokio::select! {
+                Ok(alert) = rx.recv() => {
+                    let data = serde_json::to_string(&alert).unwrap_or_default();
+                    let msg = format!("event: security_alert\ndata: {}\nretry: 3000\n\n", data);
+                    Some((Ok::<_, actix_web::Error>(actix_web::web::Bytes::from(msg)), (rx, hb, mongo)))
+                }
+                _ = hb.tick() => {
+                    Some((Ok(actix_web::web::Bytes::from_static(b": heartbeat\n\n")), (rx, hb, mongo)))
+                }
+            }
+        },
+    );
+    HttpResponse::Ok()
+        .content_type("text/event-stream")
+        .insert_header(("Cache-Control", "no-cache"))
+        .insert_header(("X-Accel-Buffering", "no"))
+        .streaming(stream)
+}
+
+// ─── Shared helpers (used by admin_ops_handlers too via super::*) ─────────────
+
+pub(crate) fn env_bool(name: &str, default: bool) -> bool {
+    match std::env::var(name) {
+        Ok(v) => matches!(
+            v.trim().to_ascii_lowercase().as_str(),
+            "1" | "true" | "yes" | "on"
+        ),
+        Err(_) => default,
+    }
+}
+
+pub(crate) async fn dns_txt_lookup(name: &str) -> Vec<String> {
+    use trust_dns_resolver::config::{ResolverConfig, ResolverOpts};
+    use trust_dns_resolver::TokioAsyncResolver;
+
+    let resolver = TokioAsyncResolver::tokio(ResolverConfig::default(), ResolverOpts::default());
+    let mut rows: Vec<String> = Vec::new();
+
+    if let Ok(lookup) = resolver.txt_lookup(name).await {
+        for txt in lookup.iter() {
+            for part in txt.txt_data() {
+                if let Ok(s) = std::str::from_utf8(part) {
+                    rows.push(s.to_string());
+                }
+            }
+        }
+    }
+
+    rows
 }
