--- src/bin/email_api_dir/monitoring_handlers/dns.rs
@@ -1,13 +1,0 @@
-use actix_web::{web, HttpResponse};
-use chrono::Utc;
-use futures_util::stream;
-use mongodb::bson::doc;
-use serde::Deserialize;
-use simple_smtp_server::monitoring;
-use simple_smtp_server::monitoring::alerts::AlertConfig;
-use simple_smtp_server::monitoring::storage;
-use simple_smtp_server::security;
-use std::sync::Arc;
-use tokio::sync::broadcast;
-
-pub use super::helpers::{dns_txt_lookup, env_bool};
