--- src/bin/email_api_dir/monitoring_handlers/mod.rs
@@ -1,12 +1,0 @@
-// monitoring_handlers module — split par domaine (cycle 16)
-
-mod helpers;
-mod types;
-mod monitoring;
-mod security;
-mod dns;
-
-pub(crate) use helpers::*;
-pub(crate) use types::*;
-pub(crate) use monitoring::*;
-pub(crate) use security::*;
